refactor: remove useless script and combine namespace
