--- Assets/Script/RoomScene.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4371427b8fa41ab4098fe4c63797cd94
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
